:sparkles: [feat] totalElements 타입 변경

+++ backend/ieum/src/main/java/com/ukcorp/ieum/chat/dto/response/ChatHistoryResponseDto.java
@@ -14,5 +14,5 @@ public class ChatHistoryResponseDto {
 
     private List<ChatHistoryDto> contents;
     private int totalPages;
-    private int totalElements;
+    private long totalElements;
 }
